refactor: MainViewController에서 Step 분리
